plugins/feed_reader/requirements.txt
feedparser
cloudscraper
PyQt6-WebEngine
pytz
python-dateutil
curl-cffi